Enable `directly-loc-start-end` for all languages
