+++ .eslintrc.yml
@@ -106,7 +106,6 @@ overrides:
     rules:
       prettier-internal-rules/prefer-fast-path-each: error
   - files:
-      - src/language-js/**/*.js
-      - src/language-css/**/*.js
+      - src/language-*/**/*.js
     rules:
       prettier-internal-rules/directly-loc-start-end: error
