+++ src/language-graphql/parser-graphql.js
@@ -2,6 +2,7 @@
 
 const createError = require("../common/parser-create-error");
 const { hasPragma } = require("./pragma");
+const { locStart, locEnd } = require("./loc");
 
 function parseComments(ast) {
   const comments = [];
@@ -77,18 +78,8 @@ module.exports = {
       parse,
       astFormat: "graphql",
       hasPragma,
-      locStart(node) {
-        if (typeof node.start === "number") {
-          return node.start;
-        }
-        return node.loc && node.loc.start;
-      },
-      locEnd(node) {
-        if (typeof node.end === "number") {
-          return node.end;
-        }
-        return node.loc && node.loc.end;
-      },
+      locStart,
+      locEnd,
     },
   },
 };
