+++ src/language-graphql/printer-graphql.js
@@ -12,6 +12,7 @@ const {
 } = require("../document").builders;
 const { hasIgnoreComment, isNextLineEmpty } = require("../common/util");
 const { insertPragma } = require("./pragma");
+const { locStart, locEnd } = require("./loc");
 
 function genericPrint(path, options, print) {
   const n = path.getValue();
@@ -31,11 +32,7 @@ function genericPrint(path, options, print) {
         if (index !== n.definitions.length - 1) {
           parts.push(hardline);
           if (
-            isNextLineEmpty(
-              options.originalText,
-              pathChild.getValue(),
-              options.locEnd
-            )
+            isNextLineEmpty(options.originalText, pathChild.getValue(), locEnd)
           ) {
             parts.push(hardline);
           }
@@ -44,7 +41,7 @@ function genericPrint(path, options, print) {
       return concat([concat(parts), hardline]);
     }
     case "OperationDefinition": {
-      const hasOperation = options.originalText[options.locStart(n)] !== "{";
+      const hasOperation = options.originalText[locStart(n)] !== "{";
       const hasName = !!n.name;
       return concat([
         hasOperation ? n.operation : "",
@@ -627,7 +624,7 @@ function printSequence(sequencePath, options, print) {
     const printed = print(path);
 
     if (
-      isNextLineEmpty(options.originalText, path.getValue(), options.locEnd) &&
+      isNextLineEmpty(options.originalText, path.getValue(), locEnd) &&
       i < count - 1
     ) {
       return concat([printed, hardline]);
