+++ src/language-markdown/printer-markdown.js
@@ -28,6 +28,7 @@ const {
 const { replaceEndOfLineWith } = require("../common/util");
 const embed = require("./embed");
 const { insertPragma } = require("./pragma");
+const { locStart, locEnd } = require("./loc");
 const preprocess = require("./preprocess");
 const clean = require("./clean");
 const {
@@ -112,7 +113,7 @@ function genericPrint(path, options, print) {
       const isFirstSentence = (node, name, index) =>
         node.type === "sentence" && index === 0;
       const isLastChildAutolink = (node, name, index) =>
-        isAutolink(node.children[index - 1], options);
+        isAutolink(node.children[index - 1]);
 
       if (
         escapedValue !== node.value &&
@@ -463,10 +464,7 @@ function genericPrint(path, options, print) {
     case "inlineMath": {
       // remark-math trims content but we don't want to remove whitespaces
       // since it's very possible that it's recognized as math accidentally
-      return options.originalText.slice(
-        options.locStart(node),
-        options.locEnd(node)
-      );
+      return options.originalText.slice(locStart(node), locEnd(node));
     }
 
     case "tableRow": // handled in "table"
